voltium: research defaults, ScoreStrategy ABC, rank-weighted book, fractional contracts
- Strategy layer: ScoreStrategy ABC (eligibility + scores -> size) shared by
  OptimizationStrategy and the new RankWeightedStrategy (centred-rank or
  quantile dollar vega, no risk model).
- TradeGeneratorConfig.fractional / .research(): unrounded contracts, no
  screens, no band. contracts is Float64 in records and Portfolio.
- Demos default to the research configuration: no costs, no screens,
  fractional, no jump penalty, net-vega + gross-vega constraints only.
  --costs, --screens, --full-constraints, --jump-penalty, --strategy rank
  restore or swap each piece. Output tag carries the strategy.
- Tests for the rank book and fractional sizing; docs updated.

+++ voltium/src/voltium/backtester.py
@@ -50,7 +50,7 @@ RECORD_SCHEMA = {
     "symbol": pl.Utf8,
     "expiration": pl.Date,
     "strike": pl.Float64,
-    "contracts": pl.Int64,
+    "contracts": pl.Float64,
     "hedge_shares": pl.Float64,
     "mid": pl.Float64,
     "underlying": pl.Float64,
